Add better dungeon parsing, fix 16 bit protocol issue, format all files

+++ zdu_ui/src/server/state.rs
@@ -2,7 +2,7 @@ use std::collections::HashMap;
 
 #[derive(Debug, Clone, Default)]
 pub struct GameState {
-    pub items: HashMap<u8, u8>,
+    pub items: HashMap<u16, u8>,
     pub change_number: u64,
 }
 
@@ -10,5 +10,5 @@ pub struct GameState {
 pub struct BroadcastUpdate {
     pub sender_id: u32,
     pub new_change_number: u64,
-    pub updates: Vec<(u8, u8)>,
+    pub updates: Vec<(u16, u8)>,
 }
